refactor: rename allow.rs -> manage_sources.rs
